eval 마무리

+++ evaluation/metric/proficiency.py
@@ -10,15 +10,15 @@ def eval_proficiency(input_conv_data, input_audio_list, input_stt_list):
 
     pause_score, pause_ratio, pause_feedback = get_pause_ratio(input_stt_list)
 
-    proficiency_score = (speed_score+pronunciation_score+pause_score)/3
+    proficiency_score = int((speed_score+pronunciation_score+pause_score)/3)
     proficiency_feedback = {
-        'speed_score': speed_score,
+        'speed_score': int(speed_score),
         'speed_spm': speed_spm,
         'speed_feedback': speed_feedback,
-        'pronunciation_score': pronunciation_score,
+        'pronunciation_score': int(pronunciation_score),
         'pronunciation_level': pronunciation_level,
         'pronunciation_feedback': pronunciation_feedback,
-        'pause_score': pause_score,
+        'pause_score': int(pause_score),
         'pause_ratio': pause_ratio,
         'pause_feedback': pause_feedback
     }
